+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/utils
 
+## 0.5.81
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.67
+
 ## 0.5.80
 
 ### Patch Changes